Feat/27 diary garden api (#49)
* feat: diary 도메인 초기 세팅

* feat: /diaries 목록 조회 api 구현, 테스트 코드 통과

* feat: /diaries 등록 post api 구현, 테스트 코드 통과

* feat: /diaries/{diaryId} 조회 get api 구현, 테스트 코드 통과

* feat: /diaries/{diaryId} 다이어리 수정 put api 구현, 테스트 코드 통과, spotless 적용

* feat: /diaries/{diaryId} 다이어리 삭제 put api 구현, 테스트 코드 통과, spotless 적용

* refactor: /diaries/{diaryId} 다이어리 수정 api 유저 인증 추가, 테스트 코드 통과

* refactor: 일기 작성, 수정에서 이미지 관련은 제거, 이미지 분리하기 위함. 테스트 코드 통과

* feat: 이미지 업로드 api 구현, s3설정 추가, aws 의존성 추가, 테스트 코드 통과

* chore: spotless 적용

* chore: spotless 적용

* chore: spotless 적용

* feat: 일기 이미지 삭제 api 구햔, DiaryImage 엔티티 추가, 테스트 코드 작성

* feat: avatarPost 엔티티 초기 세팅

* feat: 포스트 상세 조회 api 임시 구현(이미지 처리 미구현), Bookmark 엔티티 초기 구현, 테스트 작성

* feat: 오늘의 키워드 조회임시, 전체 키워드 조회 api 구현, 테스트 통과

* refactor: 날짜별 키워드 조회 api 수정.

* feat: 텃밭 물주기 api 구현, 테스트 코드 작성

* feat: 텃밭 햇빛 주기 api 구현, 테스트 코드 작성

* feat: 일기, 포스트 좋아요 누르기, 취소 api 보완, 테스트 코드 작성

* refactor: 일기, 포스트 좋아요 누르기, 취소 api 반환형 수정

* refactor: 일기, 포스트 엔티티 수정 (likeCount 반영)

* feat: 텃밭 슬롯 추가 api 구현, 테스트 코드 작성

* feat: 텃밭 배경화면 조회 api 구현, 테스트 코드 작성

* refactor: api Operation 작업

* refactor: 패키지 구조 반영

* refactor: 파일명 변경 반영

* spotless apply 적용

* spotless apply 적용

* refactor: DiaryService, AvatarPostService 보완, DiaryServiceTest 수정

* spotless apply

* refactor: 테스트 코드 오류 임시 주석

* spotless apply

* spotless apply

* refactor: register api request dto 수정

* spotless apply

* spotless apply

* refactor: import문 조정

* refactor: import문 조정

* spotless apply

* spotless apply

* refactor: LikeRepository 정리

* 테스트 코드 임시 주석화

* spotless apply

+++ src/main/java/com/example/cp_main_be/domain/member/user/dto/response/UserResponse.java
@@ -19,7 +19,7 @@ public class UserResponse {
   @Builder
   public static class LevelStatusResponseDTO {
     // GET /level 의 Response
-    private Long level; // 레벨
+    private Integer level; // 레벨
     private Long experience; // 경험
   }
 }
